marathon driver should send only configuration difference, closes #608

+++ src/main/scala/io/vamp/container_driver/marathon/api/MarathonApp.scala
@@ -7,7 +7,7 @@ case class MarathonApp(
   container: Option[Container],
   instances: Int,
   cpus: Double,
-  mem: Double,
+  mem: Int,
   env: Map[String, String],
   cmd: Option[String],
   args: List[String] = Nil,
